+++ tests/Page.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Fomvasss\MediaLibraryExtension\Tests;
+
+/**
+ * Інший тип моделі (інший morph class) на тій самій таблиці
+ */
+class Page extends Article
+{
+    protected $table = 'articles';
+}